+++ src/zm_poly.h
@@ -23,28 +23,6 @@
 #include "zm_box.h"
 #include <vector>
 
-class Edge {
- public:
-  Edge() = default;
-  Edge(int32 min_y, int32 max_y, double min_x, double _1_m) : min_y(min_y), max_y(max_y), min_x(min_x), _1_m(_1_m) {}
-
-  static bool CompareYX(const Edge &e1, const Edge &e2) {
-    if (e1.min_y == e2.min_y)
-      return e1.min_x < e2.min_x;
-    return e1.min_y < e2.min_y;
-  }
-
-  static bool CompareX(const Edge &e1, const Edge &e2) {
-    return e1.min_x < e2.min_x;
-  }
-
- public:
-  int32 min_y;
-  int32 max_y;
-  double min_x;
-  double _1_m;
-};
-
 // This class represents convex or concave non-self-intersecting polygons.
 class Polygon {
  public:
